// frontend/src/generated/models/Book.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * 書籍のタイトルおよび著者をもつクラス
 */
export type Book = {
    /**
     * id
     */
    id?: number;
    /**
     * 書籍名
     */
    title: string;
    /**
     * 著者
     */
    author: string;
};